feat: Unify data table viewers and enhance UI components
- Replace separate ParquetViewer and DataTableViewer with UniversalDataTableViewer
  - Consolidate 1,476 lines into 650 lines of optimized code
  - Implement modular component architecture with 5 subcomponents
  - Add unified data provider factory pattern

- Enhance worksheet switching interface
  - Replace dropdown with modern tab-style buttons
  - Fix border jitter issue with consistent border styling
  - Improve responsive design and visual consistency

- Optimize cell content rendering
  - Implement universal DataTableCell component with 100% type coverage
  - Add comprehensive modal support for all data types
  - Support arrays, objects, booleans, numbers, dates, and strings
  - Add intelligent expand/collapse behavior

- Fix worksheet switching functionality
  - Reset table state (sorting, filtering) on sheet change
  - Properly reload metadata and column definitions
  - Support multi-sheet Excel/ODS files correctly

- Complete internationalization support
  - Add missing English translation keys
  - Fix inconsistent translation key usage
  - Support both Chinese and English interfaces

- Add UI enhancements
  - Custom scrollbar styling for better UX
  - Improved button hover states and transitions
  - Consistent color scheme with project design
  - Enhanced accessibility and keyboard navigation

Technical improvements:
- Provider pattern for data source abstraction
- Virtual scrolling for large datasets
- Chunked loading for memory efficiency
- Type-safe component interfaces
- Comprehensive error handling
