+++ src/components/FileViewer/components/DataTableMetadata.tsx
@@ -1,0 +1,66 @@
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+import type { DataMetadata } from '../providers';
+
+interface DataTableMetadataProps {
+  metadata: DataMetadata;
+  loadedRows: number;
+  totalRows: number;
+  fileSize: number;
+}
+
+export const DataTableMetadata: React.FC<DataTableMetadataProps> = ({
+  metadata,
+  loadedRows,
+  totalRows,
+  fileSize
+}) => {
+  const { t } = useTranslation();
+
+  const formatFileSize = (bytes: number): string => {
+    if (bytes === 0) return '0 B';
+    const k = 1024;
+    const sizes = ['B', 'KB', 'MB', 'GB', 'TB'];
+    const i = Math.floor(Math.log(bytes) / Math.log(k));
+    return parseFloat((bytes / Math.pow(k, i)).toFixed(2)) + ' ' + sizes[i];
+  };
+
+  return (
+    <div className="flex-shrink-0 border-b border-gray-200 dark:border-gray-700 bg-gray-50 dark:bg-gray-800 px-4 lg:px-6 py-4">
+      <div className="grid grid-cols-2 md:grid-cols-4 gap-4 text-sm">
+        <div>
+          <span className="text-gray-600 dark:text-gray-400">
+            {t('data.table.metadata.rows')}:
+          </span>
+          <span className="ml-2 font-medium text-gray-900 dark:text-gray-100">
+            {metadata.numRows.toLocaleString()}
+          </span>
+        </div>
+        <div>
+          <span className="text-gray-600 dark:text-gray-400">
+            {t('data.table.metadata.columns')}:
+          </span>
+          <span className="ml-2 font-medium text-gray-900 dark:text-gray-100">
+            {metadata.numColumns}
+          </span>
+        </div>
+        <div>
+          <span className="text-gray-600 dark:text-gray-400">
+            {t('data.table.metadata.loaded')}:
+          </span>
+          <span className="ml-2 font-medium text-gray-900 dark:text-gray-100">
+            {loadedRows.toLocaleString()} ({((loadedRows / totalRows) * 100).toFixed(1)}%)
+          </span>
+        </div>
+        <div>
+          <span className="text-gray-600 dark:text-gray-400">
+            {t('data.table.metadata.fileSize')}:
+          </span>
+          <span className="ml-2 font-medium text-gray-900 dark:text-gray-100">
+            {formatFileSize(fileSize)}
+          </span>
+        </div>
+      </div>
+    </div>
+  );
+};
