+++ src/components/FileViewer/components/index.ts
@@ -1,0 +1,5 @@
+export { DataTableCell } from './DataTableCell';
+export { DataTableControls } from './DataTableControls';
+export { DataTableMetadata } from './DataTableMetadata';
+export { DataTableColumnPanel } from './DataTableColumnPanel';
+export { CellDetailModal } from './CellDetailModal';
