+++ src/i18n/locales/en/fileViewer.ts
@@ -87,17 +87,23 @@ export const fileViewer = {
 
   // Data Table Viewer
   'data.table.viewer': 'Data Table Viewer',
-  'data.table.loading': 'Loading data...',
+  'data.table.loading': 'Loading data file "{{fileName}}"...',
   'data.table.error': 'Error loading data',
-  'data.table.rows': 'rows',
-  'data.table.columns': 'columns',
+  'data.table.rows': 'Rows',
+  'data.table.columns': 'Columns',
+  'data.table.search.placeholder': 'Search data...',
   'data.table.search.global': 'Global search...',
   'data.table.search.column': 'Search this column...',
   'data.table.filter': 'Filter',
   'data.table.clear.filter': 'Clear filter',
-  'data.table.load.more': 'Load More',
-  'data.table.loading.more': 'Loading more...',
+  'data.table.load.more': 'Load more ({{count}} records remaining)',
+  'data.table.loading.more': 'Loading more data...',
   'data.table.loaded.rows': 'Loaded {{loaded}} / {{total}} rows',
+  'data.table.showing.filtered': 'Showing {{showing}} / {{total}} loaded records',
+  'data.table.sheet.label': 'Sheet',
+  'data.table.metadata.toggle': 'Show/Hide metadata',
+  'data.table.columns.toggle': 'Show/Hide column panel',
+  'data.table.columns.visibility': 'Column visibility',
   'data.table.all.loaded': 'All data loaded',
   'data.table.show.columns': 'Show columns',
   'data.table.hide.columns': 'Hide columns',
@@ -106,6 +112,9 @@ export const fileViewer = {
   'data.table.clear.sort': 'Clear sort',
   'data.table.cell.view.full': 'View full content',
   'data.table.cell.click.view': 'Click to view full content',
+  'data.table.cell.double.click.view': 'Double click to view details',
+  'data.table.cell.details': 'Cell Details',
+  'data.table.cell.location': 'Location: {{column}} column, row {{row}}',
   'data.table.cell.copy': 'Copy cell content',
   'data.table.modal.title': 'Full Content',
   'data.table.modal.close': 'Close',
