+++ src/i18n/locales/zh/common.ts
@@ -2,7 +2,7 @@ export const common = {
   // 应用信息
   'app.name': '数据集查看器',
   'app.initializing': '正在初始化...',
-  
+
   // 连接状态
   'connected.to': '已连接到',
 
@@ -18,6 +18,7 @@ export const common = {
   'retry': '重试',
   'loading': '加载中...',
   'optional': '(可选)',
+  'close': '关闭',
 
   // 时间相关
   'time.today': '今天',
