+++ src/i18n/locales/zh/fileViewer.ts
@@ -87,39 +87,33 @@ export const fileViewer = {
 
   // 数据表格查看器
   'data.table.viewer': '数据表格查看器',
-  'data.table.loading': '正在加载数据...',
+  'data.table.loading': '正在加载数据文件 "{{fileName}}"...',
   'data.table.error': '加载数据时出错',
   'data.table.rows': '行',
   'data.table.columns': '列',
+  'data.table.search.placeholder': '在数据中搜索...',
   'data.table.search.global': '全局搜索...',
   'data.table.search.column': '搜索此列...',
   'data.table.filter': '过滤',
   'data.table.clear.filter': '清除过滤',
-  'data.table.load.more': '加载更多',
-  'data.table.loading.more': '正在加载更多...',
+  'data.table.load.more': '加载更多 ({{count}} 条记录剩余)',
+  'data.table.loading.more': '正在加载更多数据...',
   'data.table.loaded.rows': '已加载 {{loaded}} / {{total}} 行',
-  'data.table.all.loaded': '已加载全部数据',
-  'data.table.show.columns': '显示列',
-  'data.table.hide.columns': '隐藏列',
-  'data.table.sort.asc': '升序排列',
-  'data.table.sort.desc': '降序排列',
-  'data.table.clear.sort': '清除排序',
-  'data.table.cell.view.full': '查看完整内容',
-  'data.table.cell.click.view': '点击查看完整内容',
-  'data.table.cell.copy': '复制单元格内容',
-  'data.table.modal.title': '完整内容',
-  'data.table.modal.close': '关闭',
-  'data.table.null.value': 'null',
-  'data.table.items.count': '{{count}} 项',
-  'data.table.sheet.switch': '切换工作表',
-  'data.table.sheet': '工作表',
-  'data.table.data.type': '数据类型',
-  'data.table.char.length': '字符长度',
-  'data.table.chars': '字符',
-  'data.table.array.items': 'Array ({{count}} 项)',
-  'data.table.showing.rows': '显示 {{showing}} / {{total}} 已加载行',
+  'data.table.showing.filtered': '显示 {{showing}} / {{total}} 条已加载记录',
+  'data.table.sheet.label': '工作表',
+  'data.table.metadata.toggle': '显示/隐藏元数据',
+  'data.table.columns.toggle': '显示/隐藏列控制面板',
+  'data.table.columns.visibility': '列可见性',
   'data.table.metadata.rows': '行数',
   'data.table.metadata.columns': '列数',
   'data.table.metadata.loaded': '已加载',
-  'data.table.metadata.size': '文件大小',
+  'data.table.metadata.fileSize': '文件大小',
+  'data.table.null.value': 'null',
+  'data.table.items.count': '{{count}} 项',
+  'data.table.cell.click.view': '点击查看完整内容',
+  'data.table.cell.view.full': '查看完整内容',
+  'data.table.cell.double.click.view': '双击查看详情',
+  'data.table.cell.details': '单元格详情',
+  'data.table.cell.location': '位置: {{column}} 列, 第 {{row}} 行',
+  'data.table.cell.copy': '复制内容',
 };
